# public/schema/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This the the api endpoint for AIMA Exercises",
    "version": "1.0.0",
    "title": "AIMA Exercises API"
  },
  "host": "aima-exercises.firebaseapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "Exercise",
      "description": "Access to AIMA Exercises"
    },
    {
      "name": "Ratings",
      "description": "Access Ratings of AIMA Exercises"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/question/chapter/{chapterID}/exercise/{exerciseID}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Fetch the exercise from a chapter",
        "description": "to-do",
        "operationId": "getExercise",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chapterID",
            "in": "path",
            "description": "ID of the chapter that needs to be fetched",
            "required": true,
            "type": "string",
            "minimum": 1
          },
          {
            "name": "exerciseID",
            "in": "path",
            "description": "ID of the chapter that needs to be fetched",
            "required": true,
            "type": "string",
            "minimum": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Exercise"
            }
          },
          "400": {
            "description": "Invalid Call"
          }
        }
      }
    },
    "/answer/chapter/{chapterID}/exercise/{exerciseID}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Fetch the exercise from a chapter",
        "description": "to-do",
        "operationId": "getAnswer",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chapterID",
            "in": "path",
            "description": "ID of the chapter that needs to be fetched",
            "required": true,
            "type": "string",
            "minimum": 1
          },
          {
            "name": "exerciseID",
            "in": "path",
            "description": "ID of the chapter that needs to be fetched",
            "required": true,
            "type": "string",
            "minimum": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Exercise"
            }
          },
          "400": {
            "description": "Invalid Call"
          }
        }
      }
    },
    "/rating/{chapterID}/": {
      "get": {
        "tags": [
          "Ratings"
        ],
        "summary": "Find exercises by chapter number",
        "description": "to-do get description",
        "operationId": "getRatingByChapters",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chapterID",
            "in": "path",
            "description": "ID of the chapter that contains the exercise(s)",
            "required": true,
            "type": "string",
            "minimum": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Ratings"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        }
      }
    },
    "/rating/{chapterID}/{exerciseID}": {
      "get": {
        "tags": [
          "Ratings"
        ],
        "summary": "Get exercise rating",
        "description": "to-do get description",
        "operationId": "getRatingByExercise",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chapterID",
            "in": "path",
            "description": "ID of the chapter that contains the exercise(s)",
            "required": true,
            "type": "string",
            "minimum": 1
          },
          {
            "name": "exerciseID",
            "in": "path",
            "description": "ID of the exercise that needs to be rated",
            "required": true,
            "type": "string",
            "minimum": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Ratings"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "post": {
        "tags": [
          "Ratings"
        ],
        "summary": "Update exercise rating",
        "description": "to-do delete description",
        "operationId": "postRating",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chapterID",
            "in": "path",
            "description": "ID of the chapter that contains the exercise(s)",
            "required": true,
            "type": "string",
            "minimum": 1
          },
          {
            "name": "exerciseID",
            "in": "path",
            "description": "ID of the exercise that needs to be rated",
            "required": true,
            "type": "string",
            "minimum": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Ratings"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        }
      }
    }
  },
  "definitions": {
    "Exercise": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "exercise": {
          "type": "string"
        },
        "chapter": {
          "type": "string"
        },
        "contributors": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          }
        },
        "last_modified": {
          "type": "string",
          "format": "date-time"
        },
        "content": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "description": "Order Status",
          "enum": [
            "author_created",
            "user_generated"
          ]
        },
        "complete": {
          "type": "boolean",
          "default": false
        }
      },
      "xml": {
        "name": "Exercise"
      }
    },
    "Ratings": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "exercise": {
          "type": "string"
        },
        "chapter": {
          "type": "string"
        },
        "rating": {
          "type": "integer",
          "format": "int32"
        },
        "success": {
          "type": "boolean",
          "default": false
        }
      },
      "xml": {
        "name": "Exercise"
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about AIMA Exercises",
    "url": "http://aimacode.github.io/aima-exercises"
  }
}